requirements.txt
# Django core
Django==5.1.6

# Admin interface customization
django-admin-interface==0.30.1
django-colorfield==0.14.0

# REST framework
djangorestframework==3.15.2
djangorestframework-gis==1.0

# GIS support
geopandas==1.1.1
shapely==2.1.1
fiona==1.10.1
pyproj==3.7.1

# Rich form & map widgets
django-tinymce==4.1.0
django-map-widgets==0.5.1
django-leaflet==0.32.0

# Developer tools / debugging
django-extensions==4.1
django-tables2==2.7.5
django-filter==24.3
django-widget-tweaks==1.5.0
pyinstrument==5.0.1

# Required dependencies
psycopg2-binary==2.9.10          # PostgreSQL support with PostGIS
Pillow==11.2.1                    # Image support for GIS / forms
numpy==2.3.1                      # Required by geopandas
lxml==6.0.0                       # Often needed for map widgets

# Static files
whitenoise

# Miscellaneous
python-dotenv